Provider2: Add router package
